add analog reads of board voltages, update other stuff

+++ include/inverter.hpp
@@ -55,7 +55,7 @@ public:
     bool check_inverter_enable_timeout();
     void debug_print();
     void commandRegen();
-    bool calc_and_send_current_limit(uint16_t pack_voltage, uint16_t discharge_power_limit,uint16_t charge_power_limit);
+    bool calc_and_send_current_limit(uint16_t pack_voltage, uint32_t discharge_power_limit,uint32_t charge_power_limit);
 };
 
 #endif
